device list abstraction
